requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
openai>=1.12.0
python-dotenv
